requirements.txt
flask
mysql-connector-python
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation
opentelemetry-instrumentation-flask
opentelemetry-exporter-otlp
opentelemetry-instrumentation-mysql
